dev (#66)
* Сдела профиль, еще добавлена тудушка на вывод

* откат

* сделан захардкоженный профиль

* Сделана кнопка добавления промокодов, сделана логика добавления промокодов без срока действия промокода

* Зачисление баллов по qr-коду.
Пофиксил хардкодинг в l10n.
Немного поменял привилегии.
Добавил либу для корректной миграции Enum в psql.
Исправил другие неточности.

* Сделана основная логика по промокодам

* Update create_promocode.py

* Update promocode.py

* Клавиатура выдачи прав

* menu_kb change

* добавил фильтры на привелегии (не везде)

* Refactor

* Добавил добавление/удаление категорий и товаров + магазин без торговли

* Добавил фильтры привелегий еще в пару мест + рефактор того что хочу

* privileges

* btn-cnacel -> btn_cancel

* Pr resolution

* Отформатировал

* small fixes

* Еще один фикс в дев((((((

* Reformat pep8.
Move ADMIN_CHAT_ID and SUPPORT_CHAT_ID to .env

* Reformat pep8. Added .editorconfig

* feature/edit_events123/belyakov (#58)

* middproceess

* Создание и редактирование мероприятий💪
💪
💪

* fix

* Выдача привелегий эвентов💪💪💪

* НАЧИСЛЕНИЕ БАЛЛОВ!!!!!

* FIx неймингов

* fix2

* почистил l10n

* пофиксл поддержку

* Пофиксил бдшку

* Кодстайл фиксы

* Багфиксы и оптимизации после тестирования

* unused

* Еще пару кодстайл фиксов

* Something changes

* middproceess

* Создание и редактирование мероприятий💪
💪
💪

* fix

* Выдача привелегий эвентов💪💪💪

* НАЧИСЛЕНИЕ БАЛЛОВ!!!!!

* FIx неймингов

* fix2

* почистил l10n

* Пофиксил бдшку

* Кодстайл фиксы

* Багфиксы и оптимизации после тестирования

* unused

* Еще пару кодстайл фиксов

* Something changes

* 6 REBASES!!!

* code style

* Changed locale and chat_actions.

---------


* Preserve migrations in release docker-compose.yaml

* Locale and logging changes

* Ask admins by giving points by the visiting event. And apply when visited.

!!! Needs to do migrations of the database

* Add fields for point to PromocodeActivation.py

!!! Needs to do migrations of the database

* Fixed circular import error

* Fixed error with UserFactory (> 64 symbols)

* time_zone Moscow added

* Пофиксил удаление

* callbackfactory fix

* Фикс начисления баллов и расписания

* опечтка

* dt_tz format

* dt_tz2

* Add spam protect and fix filter error

* Telegram id fix

* timezones fixed

* Add file logger. Fix cache dropping. Add single_message.py middleware.

---------

+++ telegram_bot/database/models/User.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import ForeignKey, Integer, String, Uuid
+from sqlalchemy import ForeignKey, Integer, String, Uuid, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from database.models import Base, EventAttendance, EventPrivilegeGrant, PkApply, Privilege, PromocodeActivation, Purchase
@@ -11,7 +11,7 @@ class User(Base):
 	__table_args__ = {"comment": "Таблица пользователей бота (участники, админы, модераторы)"}
 
 	id: Mapped[int] = mapped_column(Integer, primary_key=True)
-	telegram_id: Mapped[int] = mapped_column(Integer, unique=True, nullable=False, comment="Telegram id пользователя")
+	telegram_id: Mapped[int] = mapped_column(BigInteger, unique=True, nullable=False, comment="Telegram id пользователя")
 	telegram_username: Mapped[str] = mapped_column(String(255), nullable=True, comment="Telegram username пользователя")
 
 	full_name: Mapped[str] = mapped_column(String(255), nullable=False, comment="ФИО")
